flight_service/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
Pillow==9.0.0
requests==2.32.3
psycopg==3.2.3
psycopg2-binary==2.9.10
gunicorn==23.0.0